pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supermemory"
version = "0.1.0"
description = "Local-first AI memory engine with relational versioning and temporal grounding"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    { name = "Jared Goering", email = "goering.jared@gmail.com" },
]
keywords = ["ai", "memory", "embeddings", "knowledge-graph", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database :: Front-Ends",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "litellm>=1.0.0",
    "numpy>=1.24.0",
    "click>=8.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
local = [
    "sentence-transformers>=2.2.0",
]
ui = [
    # Note: UI requires npm/pnpm to be installed separately
    # Run: cd ui && npm install && npm run build
]

[project.urls]
Homepage = "https://github.com/jared-goering/supermemory"
Repository = "https://github.com/jared-goering/supermemory"
Issues = "https://github.com/jared-goering/supermemory/issues"

[project.scripts]
supermemory = "supermemory.cli:cli"

[project.optional-dependencies.dev]
# Development dependencies
ruff = ["ruff>=0.4.0"]
pytest = ["pytest>=7.0"]
pre-commit = ["pre-commit>=3.0"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "S",    # flake8-bandit (security)
]
ignore = [
    "S101",  # assert is fine in tests
    "S104",  # binding to 0.0.0.0 is intentional for local API
    "S110",  # bare except-pass is intentional in optional cache refresh
    "S310",  # urlopen is for localhost only
    "S603",  # subprocess calls are intentional
    "S607",  # partial executable paths
    "S608",  # SQL "injection" is parameterized placeholders, not actual injection
    "E402",  # late imports are intentional (lazy loading for optional deps)
    "E501",  # line length - too many long SQL/prompt strings to wrap
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools]
packages = ["supermemory"]

[tool.setuptools.package-data]
supermemory = ["py.typed"]
